Improve palm startup speed
We observed that palm was occasionally slow to run commands.
Timing functions were added to locate the slowdown, which turned out to
be caused by checking whether docker-compose is ready

- Adjusted how we check for docer compose reduced the execution time of this function from >3s to ~0.5s

Removing this check entirely is an option, since docker-compose now ships
with docker desktop.
However, this may cause unhandled exceptions for users running palm on linux